+++ Changelog.md
@@ -1,3 +1,9 @@
+### Development
+
+Enchancements:
+
+* Ability to fetch class constants
+
 ### 1.3.0 / 2019-04-09
 [full changelog](http://github.com/rspec/rspec-its/compare/v1.2.0...v1.3.0)