+++ Changelog.md
@@ -1,3 +1,6 @@
+### 1.3.1 / 2024-10-23
+[full changelog](http://github.com/rspec/rspec-its/compare/v1.3.0...v1.3.1)
+
 Bug fixes:
 
 * Prevent overridden `example` methods causing issues by creating our own